requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
python-dotenv==1.0.0
selenium==4.23.1
webdriver_manager
httpx==0.24.1
requests==2.31.0
pytest==7.4.0
pytest-asyncio==0.21.0
pymysql==1.1.1
beautifulsoup4==4.12.3
mysql-connector-python==9.0.0
openpyxl==3.1.5
tqdm==4.66.5
aiomysql==0.2.0
pandas==2.2.2
xlsxwriter==3.2.0
openai==1.51.2
python-multipart==0.0.17
translate==3.6.1
pillow==10.4.0
instagrapi==2.1.3
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
opencv-python==4.10.0.84
runwayml==2.1.4
pyperclip==1.9.0
pyautogui==0.9.54
anthropic==0.40.0
moviepy==2.1.1
rembg==2.0.62
onnxruntime==1.20.1
google-genai==1.0.0
redis==5.2.1
scikit-learn==1.7.0
torch==2.7.1
torchvision==0.22.1
python-jose==3.5.0
playwright==1.54.0
apscheduler==3.11.0
bcrypt==5.0.0
pydantic==2.10.1
email-validator==2.3.0
cryptography==45.0.5